Change to premake5 build system and reorganize
Make is not truly portable (despite the efforts of the mingw project)
and therefore a new build system was needed.  Becasue of the ease of use
and previous experience, I chose premake5.  It's not the most mature or
well integrated, but it's the one I'm most comfortable with.  It's not a
final decision and migration to another system is an option for the
future.

Additionally, because the project this was taken from required a flat
file hirearchy, a lot of reorganizing was required for the executable to
still find resource files.  The current solution (which is outlined in
the README) is hacky and will be superceded with a virtual file system
for resource loading.

NOTE: First commit message using the advice of Paul Vanderlaken.  All
      messages after this will attempt to follow his rules.

https://paulvanderlaken.com/2020/05/11/how-to-write-git-commit-message-7
steps/