Update all prints to shiny new log system
During this I also found a bug in the logger.  For va_lists, a copy must
be created for each consumer of it.  In this instance, both the
TermLoggerOutput and FileLoggerOutput were using the same list and it
was exhausted before it reached FileLoggerOutput (causing a segfault).
As the lists can't be rewound, a new one has to be created (copied) for
each call to LoggerOutput::output().